Выполнил 4 задание
