+++ src/create-element.js
@@ -1,0 +1,6 @@
+export default (template) => {
+  const element = document.createElement(`template`);
+  element.innerHTML = template;
+
+  return element.content.firstChild;
+};
